python_practices/fa_examp/fa_auth_project/requirements.txt
fastapi
uvicorn
bcrypt
python-jose
aiofiles
jinja2
passlib
sqlalchemy
aiosqlite
